classify store mutations
