+++ src/store/mutations/comment.js
@@ -1,0 +1,20 @@
+import _ from 'lodash'
+
+const SET_COMMENT_COUNT = (state, { count, postId, }) => {
+  let commentCount = _.find(_.get(state, [ 'commentCount', ]), { postId, })
+  if (!commentCount) {
+    commentCount = { postId, count: 0, }
+    _.get(state, [ 'commentCount', ]).push(commentCount)
+  }
+  commentCount.count = count || 0
+}
+
+const SET_COMMENTS_ME = (state, { comments, }) => {
+  const profile = state['profile']
+  profile && (profile.comments = comments)
+}
+
+export {
+  SET_COMMENT_COUNT,
+  SET_COMMENTS_ME,
+}
